Fix json imports, move json to util
