+++ packages/dds/tree/src/change-family/fence.json
@@ -1,5 +1,5 @@
 {
     "tags": [ "change-family" ],
     "exports": [ "index" ],
-    "imports": [ "rebase", "tree" ]
+    "imports": [ "rebase", "tree", "util" ]
 }
