+++ packages/dds/tree/src/feature-libraries/modular-schema/modularChangeFamily.ts
@@ -4,11 +4,11 @@
  */
 
 import { assert } from "@fluidframework/common-utils";
-import { ChangeEncoder, ChangeFamily, JsonCompatibleReadOnly, ProgressiveEditBuilder } from "../../change-family";
+import { ChangeEncoder, ChangeFamily, ProgressiveEditBuilder } from "../../change-family";
 import { ChangeRebaser } from "../../rebase";
 import { FieldKindIdentifier } from "../../schema-stored";
 import { AnchorSet, Delta, FieldKey } from "../../tree";
-import { brand, getOrAddEmptyToMap } from "../../util";
+import { brand, getOrAddEmptyToMap, JsonCompatibleReadOnly } from "../../util";
 import { FieldChangeHandler, FieldChangeMap, FieldChange } from "./fieldChangeHandler";
 import { FieldKind } from "./fieldKind";
